more file renames, directory renames
