+++ DataFormats/Detectors/FIT/CMakeLists.txt
@@ -8,5 +8,5 @@
 # granted to it by virtue of its status as an Intergovernmental Organization or
 # submit itself to any jurisdiction.
 
-add_subdirectory(T0)
+add_subdirectory(FT0)
 add_subdirectory(V0)
