+++ DataFormats/Detectors/FIT/FT0/include/DataFormatsFT0/HitType.h
