+++ DataFormats/Detectors/FIT/FT0/src/RecPoints.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "DataFormatsFITT0/RecPoints.h"
+#include "DataFormatsFT0/RecPoints.h"
 #include <cmath>
 #include <cassert>
 #include <iostream>
