+++ Detectors/FIT/FT0/simulation/CMakeLists.txt
@@ -11,6 +11,6 @@
 o2_add_library(FT0Simulation
                SOURCES src/Detector.cxx
                PUBLIC_LINK_LIBRARIES O2::SimulationDataFormat O2::FT0Base
-                                     O2::DataFormatsFITT0)
+                                     O2::DataFormatsFT0)
 
 o2_target_root_dictionary(FT0Simulation HEADERS include/FT0Simulation/Detector.h)
