+++ Detectors/FIT/FT0/simulation/include/FT0Simulation/Detector.h
@@ -17,7 +17,7 @@
 #include "SimulationDataFormat/BaseHits.h"
 #include "DetectorsBase/Detector.h" // for Detector
 #include "FT0Base/Geometry.h"
-#include "DataFormatsFITT0/HitType.h"
+#include "DataFormatsFT0/HitType.h"
 
 class FairModule;
