+++ Detectors/FIT/common/simulation/CMakeLists.txt
@@ -10,7 +10,7 @@
 
 o2_add_library(FITSimulation
                SOURCES src/Digitizer.cxx
-               PUBLIC_LINK_LIBRARIES O2::CommonDataFormat O2::DataFormatsFITT0
+               PUBLIC_LINK_LIBRARIES O2::CommonDataFormat O2::DataFormatsFT0
                                      O2::FT0Simulation)
 
 o2_target_root_dictionary(
