+++ Detectors/FIT/macros/CMakeLists.txt
@@ -9,7 +9,7 @@
 # submit itself to any jurisdiction.
 
 o2_add_test_root_macro(readHitsDigits.C
-                       PUBLIC_LINK_LIBRARIES O2::DataFormatsFITT0
+                       PUBLIC_LINK_LIBRARIES O2::DataFormatsFT0
                        LABELS fit)
 
 o2_add_test_root_macro(run_reco_t0.C
