+++ Detectors/FIT/workflow/include/FITWorkflow/FT0RecPointReaderSpec.h
@@ -17,7 +17,7 @@
 
 #include "Framework/DataProcessorSpec.h"
 #include "Framework/Task.h"
-#include "DataFormatsFITT0/RecPoints.h"
+#include "DataFormatsFT0/RecPoints.h"
 
 using namespace o2::framework;
